requirements.txt
livekit-agents
livekit-plugins-deepgram
livekit-plugins-openai
livekit-plugins-cartesia
livekit-plugins-silero
livekit-plugins-turn-detector
livekit-plugins-elevenlabs
python-dotenv